+++ sanic/cookies/__init__.py
@@ -1,0 +1,3 @@
+from .response import Cookie, CookieJar
+
+__all__ = ("Cookie", "CookieJar")
