+++ sanic/request/types.py
@@ -8,7 +8,6 @@ from typing import (
     DefaultDict,
     Dict,
     List,
-    NamedTuple,
     Optional,
     Tuple,
     Union,
@@ -22,18 +21,14 @@ from sanic.http.stream import Stream
 from sanic.models.asgi import ASGIScope
 from sanic.models.http_types import Credentials
 
-
 if TYPE_CHECKING:
     from sanic.server import ConnInfo
     from sanic.app import Sanic
 
-import email.utils
-import unicodedata
 import uuid
-
 from collections import defaultdict
 from types import SimpleNamespace
-from urllib.parse import parse_qs, parse_qsl, unquote, urlunparse
+from urllib.parse import parse_qs, parse_qsl, urlunparse
 
 from httptools import parse_url
 from httptools.parser.errors import HttpParserInvalidURLError
@@ -45,7 +40,7 @@ from sanic.constants import (
     IDEMPOTENT_HTTP_METHODS,
     SAFE_HTTP_METHODS,
 )
-from sanic.cookies import parse_cookie
+from sanic.cookies.request import CookieRequestParameters, parse_cookie
 from sanic.exceptions import BadRequest, BadURL, ServerError
 from sanic.headers import (
     AcceptList,
@@ -58,49 +53,17 @@ from sanic.headers import (
     parse_xforwarded,
 )
 from sanic.http import Stage
-from sanic.log import deprecation, error_logger, logger
+from sanic.log import deprecation, error_logger
 from sanic.models.protocol_types import TransportProtocol
 from sanic.response import BaseHTTPResponse, HTTPResponse
 
+from .form import parse_multipart_form
+from .parameters import RequestParameters
 
 try:
     from ujson import loads as json_loads  # type: ignore
 except ImportError:
     from json import loads as json_loads  # type: ignore
-
-
-class RequestParameters(dict):
-    """
-    Hosts a dict with lists as values where get returns the first
-    value of the list and getlist returns the whole shebang
-    """
-
-    def get(self, name: str, default: Optional[Any] = None) -> Optional[Any]:
-        """Return the first value, either the default or actual"""
-        return super().get(name, [default])[0]
-
-    def getlist(
-        self, name: str, default: Optional[Any] = None
-    ) -> Optional[Any]:
-        """
-        Return the entire list
-        """
-        return super().get(name, default)
-
-
-class CookieRequestParameters(RequestParameters):
-    def __getitem__(self, key: str) -> Optional[str]:
-        deprecation(
-            f"You are accessing cookie key '{key}', which is currently in "
-            "compat mode returning a single cookie value. Starting in v23.9 "
-            "accessing a cookie value like this will return a list of values. "
-            "To avoid this behavior and continue accessing a single value, "
-            f"please upgrade from request.cookies['{key}'] to "
-            f"request.cookies.get('{key}'). See more details: ___.",
-            23.9,
-        )
-        value = super().__getitem__(key)
-        return value[0]
 
 
 class Request:
@@ -1040,101 +1003,3 @@ class Request:
         :rtype: bool
         """
         return self.method in CACHEABLE_HTTP_METHODS
-
-
-class File(NamedTuple):
-    """
-    Model for defining a file. It is a ``namedtuple``, therefore you can
-    iterate over the object, or access the parameters by name.
-
-    :param type: The mimetype, defaults to text/plain
-    :param body: Bytes of the file
-    :param name: The filename
-    """
-
-    type: str
-    body: bytes
-    name: str
-
-
-def parse_multipart_form(body, boundary):
-    """
-    Parse a request body and returns fields and files
-
-    :param body: bytes request body
-    :param boundary: bytes multipart boundary
-    :return: fields (RequestParameters), files (RequestParameters)
-    """
-    files = {}
-    fields = {}
-
-    form_parts = body.split(boundary)
-    for form_part in form_parts[1:-1]:
-        file_name = None
-        content_type = "text/plain"
-        content_charset = "utf-8"
-        field_name = None
-        line_index = 2
-        line_end_index = 0
-        while not line_end_index == -1:
-            line_end_index = form_part.find(b"\r\n", line_index)
-            form_line = form_part[line_index:line_end_index].decode("utf-8")
-            line_index = line_end_index + 2
-
-            if not form_line:
-                break
-
-            colon_index = form_line.index(":")
-            idx = colon_index + 2
-            form_header_field = form_line[0:colon_index].lower()
-            form_header_value, form_parameters = parse_content_header(
-                form_line[idx:]
-            )
-
-            if form_header_field == "content-disposition":
-                field_name = form_parameters.get("name")
-                file_name = form_parameters.get("filename")
-
-                # non-ASCII filenames in RFC2231, "filename*" format
-                if file_name is None and form_parameters.get("filename*"):
-                    encoding, _, value = email.utils.decode_rfc2231(
-                        form_parameters["filename*"]
-                    )
-                    file_name = unquote(value, encoding=encoding)
-
-                # Normalize to NFC (Apple MacOS/iOS send NFD)
-                # Notes:
-                # - No effect for Windows, Linux or Android clients which
-                #   already send NFC
-                # - Python open() is tricky (creates files in NFC no matter
-                #   which form you use)
-                if file_name is not None:
-                    file_name = unicodedata.normalize("NFC", file_name)
-
-            elif form_header_field == "content-type":
-                content_type = form_header_value
-                content_charset = form_parameters.get("charset", "utf-8")
-
-        if field_name:
-            post_data = form_part[line_index:-4]
-            if file_name is None:
-                value = post_data.decode(content_charset)
-                if field_name in fields:
-                    fields[field_name].append(value)
-                else:
-                    fields[field_name] = [value]
-            else:
-                form_file = File(
-                    type=content_type, name=file_name, body=post_data
-                )
-                if field_name in files:
-                    files[field_name].append(form_file)
-                else:
-                    files[field_name] = [form_file]
-        else:
-            logger.debug(
-                "Form-data field does not have a 'name' parameter "
-                "in the Content-Disposition header"
-            )
-
-    return RequestParameters(fields), RequestParameters(files)
